requirements.txt
python-dotenv
ibm-watson
pydub